기능정리 및 package refactor
