+++ docs/README.md
@@ -1,0 +1,27 @@
+## 기능 구현 목록
+
+### Front(입출력 화면)
+- [ ] InputView
+  - [ ] 다리 길이 입력 화면
+  - [ ] 이동할 칸 선택 화면 화면([String] U, D)
+  - [ ] 재시도 여부 선택 화면([String] R, Q)
+- [ ] OutputView
+  - [ ] 이동 후 결과 출력 화면
+  - [ ] 최종 성공여부/시도 횟수 출력 화면
+
+### Back
+- [ ] 다리 생성 기능
+    - [ ] [예외]
+        - [ ] Input: 3 ~ 20 숫자 범위 벗어나면 예외
+- [ ] 다리 이동 기능
+  - [ ] [예외]
+    - [ ] Input: U 혹은 D 아니면 예외
+  - [ ] 이동 실패시 게임 reset -> 종료/재시도
+- [ ] 게임 재시도 기능
+  - [ ] [예외]
+    - [ ] Input: R 혹은 Q 아니면 예외
+
+
+### 유닛테스트
+- [ ] Input Test
+  - [ ] 예외적용여부
